pref(utils): pref utils file
