+++ src/Pipline/Grid.tsx
@@ -10,7 +10,7 @@ import { mergePipelinesWithResponse } from '@root/help'
 import Pipeline from '@root/Pipline/Pipeline'
 import Titan from '@root/Titan/Titan'
 import Auth from '@root/Auth/Auth'
-import { getLayouts, saveLayouts } from '../Utils/LayoutStorage'
+import { getLayouts, saveLayouts } from '../Utils/LayoutStorageUtils'
 import { DEFAULT_ITEM_LAYOUT } from '../Constants/Grid'
 import { PIPELINE_AUTO_REFRESH_PERIOD } from '../Constants/Config'
